Refactor: SRP, modularization
